Remove Onelogin namespace
