--- lib/onelogin/ruby-saml.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-require 'onelogin/ruby-saml/logging'
-require 'onelogin/ruby-saml/saml_message'
-require 'onelogin/ruby-saml/authrequest'
-require 'onelogin/ruby-saml/logoutrequest'
-require 'onelogin/ruby-saml/logoutresponse'
-require 'onelogin/ruby-saml/attributes'
-require 'onelogin/ruby-saml/slo_logoutrequest'
-require 'onelogin/ruby-saml/slo_logoutresponse'
-require 'onelogin/ruby-saml/response'
-require 'onelogin/ruby-saml/settings'
-require 'onelogin/ruby-saml/attribute_service'
-require 'onelogin/ruby-saml/http_error'
-require 'onelogin/ruby-saml/validation_error'
-require 'onelogin/ruby-saml/metadata'
-require 'onelogin/ruby-saml/idp_metadata_parser'
-require 'onelogin/ruby-saml/utils'
-require 'onelogin/ruby-saml/version'
