--- lib/onelogin/ruby-saml/http_error.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module OneLogin
-  module RubySaml
-    class HttpError < StandardError
-    end
-  end
-end
-
