--- lib/onelogin/ruby-saml/setting_error.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module OneLogin
-  module RubySaml
-    class SettingError < StandardError
-    end
-  end
-end
