--- lib/onelogin/ruby-saml/validation_error.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module OneLogin
-  module RubySaml
-    class ValidationError < StandardError
-    end
-  end
-end
-
