--- lib/onelogin/ruby-saml/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module OneLogin
-  module RubySaml
-    VERSION = '1.15.0'
-  end
-end
