+++ lib/ruby-saml.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require 'onelogin/ruby-saml'
+require 'ruby_saml'
