+++ lib/ruby_saml.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require 'ruby_saml/logging'
+require 'ruby_saml/saml_message'
+require 'ruby_saml/authrequest'
+require 'ruby_saml/logoutrequest'
+require 'ruby_saml/logoutresponse'
+require 'ruby_saml/attributes'
+require 'ruby_saml/slo_logoutrequest'
+require 'ruby_saml/slo_logoutresponse'
+require 'ruby_saml/response'
+require 'ruby_saml/settings'
+require 'ruby_saml/attribute_service'
+require 'ruby_saml/http_error'
+require 'ruby_saml/validation_error'
+require 'ruby_saml/metadata'
+require 'ruby_saml/idp_metadata_parser'
+require 'ruby_saml/utils'
+require 'ruby_saml/version'
+
+# For backward compatibility
+module OneLogin
+end
+OneLogin::RubySaml = RubySaml
