+++ lib/ruby_saml/http_error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module RubySaml
+  class HttpError < StandardError
+  end
+end
