+++ lib/ruby_saml/setting_error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module RubySaml
+  class SettingError < StandardError
+  end
+end
