+++ lib/ruby_saml/validation_error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module RubySaml
+  class ValidationError < StandardError
+  end
+end
