+++ lib/ruby_saml/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module RubySaml
+  VERSION = '2.0.0'
+end
