+++ lib/xml_security.rb
@@ -31,8 +31,8 @@ require "openssl"
 require 'nokogiri'
 require "digest/sha1"
 require "digest/sha2"
-require "onelogin/ruby-saml/utils"
-require "onelogin/ruby-saml/error_handling"
+require "ruby_saml/utils"
+require "ruby_saml/error_handling"
 
 module XMLSecurity
 
@@ -185,7 +185,7 @@ module XMLSecurity
   end
 
   class SignedDocument < BaseDocument
-    include OneLogin::RubySaml::ErrorHandling
+    include RubySaml::ErrorHandling
 
     attr_writer :signed_element_id
 
@@ -207,7 +207,7 @@ module XMLSecurity
       )
 
       if cert_element
-        base64_cert = OneLogin::RubySaml::Utils.element_text(cert_element)
+        base64_cert = RubySaml::Utils.element_text(cert_element)
         cert_text = Base64.decode64(base64_cert)
         begin
           cert = OpenSSL::X509::Certificate.new(cert_text)
@@ -245,7 +245,7 @@ module XMLSecurity
       )
 
       if cert_element
-        base64_cert = OneLogin::RubySaml::Utils.element_text(cert_element)
+        base64_cert = RubySaml::Utils.element_text(cert_element)
         cert_text = Base64.decode64(base64_cert)
         begin
           cert = OpenSSL::X509::Certificate.new(cert_text)
@@ -293,7 +293,7 @@ module XMLSecurity
         "./ds:SignatureValue",
         {"ds" => DSIG}
       )
-      signature = Base64.decode64(OneLogin::RubySaml::Utils.element_text(base64_signature))
+      signature = Base64.decode64(RubySaml::Utils.element_text(base64_signature))
 
       # canonicalization method
       canon_algorithm = canon_algorithm REXML::XPath.first(
@@ -337,7 +337,7 @@ module XMLSecurity
         "//ds:DigestValue",
         { "ds" => DSIG }
       )
-      digest_value = Base64.decode64(OneLogin::RubySaml::Utils.element_text(encoded_digest_value))
+      digest_value = Base64.decode64(RubySaml::Utils.element_text(encoded_digest_value))
 
       unless digests_match?(hash, digest_value)
         return append_error("Digest mismatch", soft)
