+++ test/logout_responses/logoutresponse_fixtures.rb
@@ -1,10 +1,10 @@
-#encoding: utf-8
+# frozen_string_literal: true
 
 def default_logout_response_opts
   {
-      :uuid => "_28024690-000e-0130-b6d2-38f6b112be8b",
-      :issue_instant => Time.now.strftime('%Y-%m-%dT%H:%M:%SZ'),
-      :settings => settings
+    uuid: "_28024690-000e-0130-b6d2-38f6b112be8b",
+    issue_instant: Time.now.strftime('%Y-%m-%dT%H:%M:%SZ'),
+    settings: settings
   }
 end
 
@@ -71,16 +71,14 @@ def invalid_xml_logout_response_document
 end
 
 def settings
-  @settings ||= OneLogin::RubySaml::Settings.new(
-      {
-          :assertion_consumer_service_url => "http://app.muda.no/sso/consume",
-          :single_logout_service_url => "http://app.muda.no/sso/consume_logout",
-          :issuer => "http://app.muda.no",
-          :sp_name_qualifier => "http://sso.muda.no",
-          :idp_sso_service_url => "http://sso.muda.no/sso",
-          :idp_slo_service_url => "http://sso.muda.no/slo",
-          :idp_cert_fingerprint => "00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00",
-          :name_identifier_format => "urn:oasis:names:tc:SAML:2.0:nameid-format:transient",
-      }
-  )
+  @settings ||= RubySaml::Settings.new({
+    assertion_consumer_service_url: "http://app.muda.no/sso/consume",
+    single_logout_service_url: "http://app.muda.no/sso/consume_logout",
+    issuer: "http://app.muda.no",
+    sp_name_qualifier: "http://sso.muda.no",
+    idp_sso_service_url: "http://sso.muda.no/sso",
+    idp_slo_service_url: "http://sso.muda.no/slo",
+    idp_cert_fingerprint: "00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00",
+    name_identifier_format: "urn:oasis:names:tc:SAML:2.0:nameid-format:transient"
+  })
 end
