+++ test/logoutrequest_test.rb
@@ -1,6 +1,6 @@
 require_relative 'test_helper'
 
-require 'onelogin/ruby-saml/logoutrequest'
+require 'ruby_saml/logoutrequest'
 
 class RequestTest < Minitest::Test
 
@@ -102,10 +102,10 @@ class RequestTest < Minitest::Test
       end
 
       it "creates request with ID is prefixed, when :id_prefix is passed" do
-        OneLogin::RubySaml::Utils::set_prefix("test")
+        RubySaml::Utils.set_prefix("test")
         request = OneLogin::RubySaml::Logoutrequest.new
         assert_match(/^test/, request.uuid)
-        OneLogin::RubySaml::Utils::set_prefix("_")
+        RubySaml::Utils.set_prefix("_")
       end
     end
